requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3
cx-Oracle==6.4.1
django-adminlte2==0.4.1
graphene==2.1.3
graphene-django==2.1.0
django-graphiql
Pillow==5.2.0
promise==2.1
psycopg2==2.7.5
pycparser==2.18
PyMySQL==0.9.2
typing==3.6.4
django-graphql-jwt==0.2.0
